Make uuid_generator more robust
We are currently extracting the filename using split, but it won't catch
weird edge cases such as .tar.gz. Python has builtin tools for this
which we will use, as well as normalizing the extension just in case.
(For example, design gives images to us in form of .JPEG). Also changed
the f string because ruff wanted me to lol.